+++ aa_bulletin_board/__init__.py
@@ -5,6 +5,6 @@ App init
 # Django
 from django.utils.translation import gettext_lazy as _
 
-__version__ = "2.5.0"
+__version__ = "2.5.1"
 __title__ = "Bulletin Board"
 __title_translated__ = _("Bulletin Board")