+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Section Order:
 
 <!-- Your changes go here -->
 
+## [2.5.1] - 2025-12-02
+
 ### Removed
 
 - Unnecessary URL parameter
@@ -1112,6 +1114,7 @@ python manage.py migrate
 [2.4.3]: https://github.com/ppfeufer/aa-bulletin-board/compare/v2.4.2...v2.4.3 "v2.4.3"
 [2.4.4]: https://github.com/ppfeufer/aa-bulletin-board/compare/v2.4.3...v2.4.4 "v2.4.4"
 [2.5.0]: https://github.com/ppfeufer/aa-bulletin-board/compare/v2.4.4...v2.5.0 "v2.5.0"
-[in development]: https://github.com/ppfeufer/aa-bulletin-board/compare/v2.5.0...HEAD "In Development"
+[2.5.1]: https://github.com/ppfeufer/aa-bulletin-board/compare/v2.5.0...v2.5.1 "v2.5.1"
+[in development]: https://github.com/ppfeufer/aa-bulletin-board/compare/v2.5.1...HEAD "In Development"
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a Changelog"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic Versioning"